pyproject.toml
[tool.poetry]
name = "kr-oop"
version = "0.1.0"
description = ""
authors = ["Anton Glazunov <tagl91@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
